sql: consolidate COMMENT ON tree files
Merge the ten per-object comment_on_*.go files into a single comment_on.go.
Add parse-roundtrip test coverage for the COMMENT ON variants that were
previously missing from pkg/sql/parser/testdata/comment.

Release note: None
