+++ pkg/sql/sem/tree/BUILD.bazel
@@ -31,16 +31,7 @@ go_library(
         "changefeed.go",
         "check.go",
         "col_name.go",
-        "comment_on_column.go",
-        "comment_on_constraint.go",
-        "comment_on_database.go",
-        "comment_on_index.go",
-        "comment_on_routine.go",
-        "comment_on_schema.go",
-        "comment_on_sequence.go",
-        "comment_on_table.go",
-        "comment_on_type.go",
-        "comment_on_view.go",
+        "comment_on.go",
         "compare.go",
         "constant.go",
         "constant_eval.go",
