--- pkg/sql/sem/tree/comment_on_database.go
@@ -1,32 +1,0 @@
-// Copyright 2018 The Cockroach Authors.
-//
-// Use of this software is governed by the CockroachDB Software License
-// included in the /LICENSE file.
-
-package tree
-
-import "github.com/cockroachdb/cockroach/pkg/sql/lexbase"
-
-// CommentOnDatabase represents an COMMENT ON DATABASE statement.
-type CommentOnDatabase struct {
-	Name    Name
-	Comment *string
-}
-
-// Format implements the NodeFormatter interface.
-func (n *CommentOnDatabase) Format(ctx *FmtCtx) {
-	ctx.WriteString("COMMENT ON DATABASE ")
-	ctx.FormatNode(&n.Name)
-	ctx.WriteString(" IS ")
-	if n.Comment != nil {
-		// TODO(knz): Replace all this with ctx.FormatNode
-		// when COMMENT supports expressions.
-		if ctx.flags.HasFlags(FmtHideConstants) {
-			ctx.WriteString("'_'")
-		} else {
-			lexbase.EncodeSQLStringWithFlags(&ctx.Buffer, *n.Comment, ctx.flags.EncodeFlags())
-		}
-	} else {
-		ctx.WriteString("NULL")
-	}
-}
