--- pkg/sql/sem/tree/comment_on_index.go
@@ -1,32 +1,0 @@
-// Copyright 2019 The Cockroach Authors.
-//
-// Use of this software is governed by the CockroachDB Software License
-// included in the /LICENSE file.
-
-package tree
-
-import "github.com/cockroachdb/cockroach/pkg/sql/lexbase"
-
-// CommentOnIndex represents a COMMENT ON INDEX statement.
-type CommentOnIndex struct {
-	Index   TableIndexName
-	Comment *string
-}
-
-// Format implements the NodeFormatter interface.
-func (n *CommentOnIndex) Format(ctx *FmtCtx) {
-	ctx.WriteString("COMMENT ON INDEX ")
-	ctx.FormatNode(&n.Index)
-	ctx.WriteString(" IS ")
-	if n.Comment != nil {
-		// TODO(knz): Replace all this with ctx.FormatNode
-		// when COMMENT supports expressions.
-		if ctx.flags.HasFlags(FmtHideConstants) {
-			ctx.WriteString("'_'")
-		} else {
-			lexbase.EncodeSQLStringWithFlags(&ctx.Buffer, *n.Comment, ctx.flags.EncodeFlags())
-		}
-	} else {
-		ctx.WriteString("NULL")
-	}
-}
