--- pkg/sql/sem/tree/comment_on_schema.go
@@ -1,32 +1,0 @@
-// Copyright 2021 The Cockroach Authors.
-//
-// Use of this software is governed by the CockroachDB Software License
-// included in the /LICENSE file.
-
-package tree
-
-import "github.com/cockroachdb/cockroach/pkg/sql/lexbase"
-
-// CommentOnSchema represents an COMMENT ON SCHEMA statement.
-type CommentOnSchema struct {
-	Name    ObjectNamePrefix
-	Comment *string
-}
-
-// Format implements the NodeFormatter interface.
-func (n *CommentOnSchema) Format(ctx *FmtCtx) {
-	ctx.WriteString("COMMENT ON SCHEMA ")
-	ctx.FormatNode(&n.Name)
-	ctx.WriteString(" IS ")
-	if n.Comment != nil {
-		// TODO(knz): Replace all this with ctx.FormatNode
-		// when COMMENT supports expressions.
-		if ctx.flags.HasFlags(FmtHideConstants) {
-			ctx.WriteByte('_')
-		} else {
-			lexbase.EncodeSQLStringWithFlags(&ctx.Buffer, *n.Comment, ctx.flags.EncodeFlags())
-		}
-	} else {
-		ctx.WriteString("NULL")
-	}
-}
